% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndecomp.R
\name{bndco}
\alias{bndco}
\title{Beveridge & Nelson Decomposition}
\format{\describe{
  \item{}{log(RGDP)}
}}
\source{
Example data (rlgdp):log series of U.S Real GDP (1947:1 to 2008:2)
 dowloaded from \url{http://time-series.net/data_sets}
}
\usage{
bndco(mod_data, pm = 2, qm = 2)
}
\arguments{
\item{y:}{the series to decompose}

\item{pm:}{maximum lag for AR (default=2)}

\item{qm:}{maximum lag for MA (default=2)}
}
\value{
bnd returns a 2-column matrix of, respectively, the permanent and cyclical components in y
}
\description{
Implements Beveridge and Nelson decomposition of series into a permanent
component and a cyclical one.
}
\examples{
bndco(lrgdp,1,1)   # Quarterly US log-GDP(1947:2008) --> arima(1,1,0) data
y <- rnorm(1:100)   # Level-Stationary --> Stop
y <- cumsum(rnorm(1:100)) + (1:100)**3   # Non-Stationary 1st difference --> Stop

}
\author{
Deepankar Basu, \email{dbasu@econs.umass.edu}

Fouad Ferdi, \email{fouad.ferdi@univ-paris13.fr}
}
