% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_decomposition.R
\name{logit_decomposition}
\alias{logit_decomposition}
\title{Logit Decomposition}
\usage{
logit_decomposition(.data, y, x, independent = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{
A tibble with the with the contribution of each covariate to the change of the structural coefficient.
}
\description{
Logit Decomposition
}
